peaceful end

+++ src/main/java/ir/sobhan/restapi/dao/CourseSectionRegistrationRepository.java
@@ -3,5 +3,5 @@ package ir.sobhan.restapi.dao;
 import ir.sobhan.restapi.model.coursesection.CourseSectionRegistration;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseSectionRegistrationRepository extends JpaRepository<CourseSectionRegistration, Integer> {
+public interface CourseSectionRegistrationRepository extends JpaRepository<CourseSectionRegistration, Long> {
 }
