+++ src/main/java/ir/sobhan/restapi/dao/CustomUserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CustomUserRepository extends JpaRepository<CustomUser, Integer> {
+public interface CustomUserRepository extends JpaRepository<CustomUser, Long> {
     Optional<CustomUser> findByUsername(String username);
 }
