+++ src/main/java/ir/sobhan/restapi/dao/InstructorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InstructorRepository extends JpaRepository<Instructor, Integer> {
+public interface InstructorRepository extends JpaRepository<Instructor, Long> {
     Optional<Instructor> findByCustomUserUsername(String username);
 }
