+++ src/main/java/ir/sobhan/restapi/dao/StaffRepository.java
@@ -3,5 +3,5 @@ package ir.sobhan.restapi.dao;
 import ir.sobhan.restapi.model.individual.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffRepository extends JpaRepository<Staff, Integer> {
+public interface StaffRepository extends JpaRepository<Staff, Long> {
 }
