+++ src/main/java/ir/sobhan/restapi/dao/TermRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TermRepository extends JpaRepository<Term, Integer> {
+public interface TermRepository extends JpaRepository<Term, Long> {
     Optional<Term> findByTitle(String title);
     void deleteTermByTitle(String title);
 }
