+++ src/main/java/ir/sobhan/restapi/dao/TokenRepository.java
@@ -3,11 +3,11 @@ package ir.sobhan.restapi.dao;
 import java.util.List;
 import java.util.Optional;
 
-import ir.sobhan.restapi.auth.token.Token;
+import ir.sobhan.restapi.auth.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     @Query(value = """
       select t from Token t inner join CustomUser u\s
